Add support for dimod~=0.10.0
Note: it's safe to omit dtype in bqm.to_numpy_vectors() as in
dimod 0.9 it'll default to float64 anyway, and in dimod 0.10 it'll be
inferred from data.

Beware: index type is int64 in dimod 0.10, yet we still use int in C code.
Plan is to keep it like this until we need to support more than 2B variables
in greedy, and/or an efficient bqm data interface is made available.