Fix CI python env caching
Don't cache python env anymore. The env references the python binary
which can change on image update (e.g. python patch release). This
breaks the ref, so we would either need to cache the pyenv's python
binary, or include the full python version in the cache key. All this
makes is not worth the few seconds saved on a fresh env build. And not
caching is consistent with other jobs.